requirements.txt
# Core backend and API
fastapi==0.104.1
uvicorn==0.24.0

# File parsing and data handling
PyMuPDF
pdfplumber
pandas==2.0.3
openpyxl==3.1.2

# Image processing
opencv-python
Pillow==10.0.1

# OCR
google-cloud-vision

# AI and embedding
openai==1.3.7
weaviate-client==3.25.3

# Frontend
streamlit==1.47.1

# Speech recognition
SpeechRecognition==3.14.3
pyaudio==0.2.14

# Utility
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0

# Database
pymongo==4.6.0
psycopg2-binary==2.9.9

# Authentication
firebase-admin==6.2.0
pyjwt==2.8.0

# PDF Generation
PyPDF2==3.0.1
reportlab==4.0.7

# Email
sendgrid==6.10.0

# Analytics and Visualization
plotly==6.2.0
numpy==1.24.3

# Additional utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
